% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_betainact}
\alias{c_betainact}
\title{find element from one vector in another vector}
\usage{
c_betainact(beta, active)
}
\value{
the target vector elements
}
\description{
find element from one vector in another vector
}
