% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_betaSD}
\alias{c_betaSD}
\title{re-standardize the centered X matrix}
\usage{
c_betaSD(Xstar, groups, beta, Qmat, Dvec, G, n)
}
\arguments{
\item{Xstar}{centered numeric matrix}

\item{beta}{estimated parameters}

\item{Qmat, Dvec}{SVD results}

\item{G}{number of groups}

\item{n}{length of subjects}

\item{group}{vector group index}
}
\value{
A standardized matrix
}
\description{
re-standardize the centered X matrix
}
