% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_Xtilde}
\alias{c_Xtilde}
\title{Parametrize X matrix}
\usage{
c_Xtilde(Xstar, groups, G, n)
}
\arguments{
\item{Xstar}{input centered matrix}

\item{groups}{group index}

\item{G}{the number of unique groups}

\item{n}{length of outcome variable}
}
\value{
A list with output matrix and SVD calculation
}
\description{
Parametrize X matrix
}
