% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSGL.R
\name{SSGL}
\alias{SSGL}
\title{A function to pergorm ssgl algorithm}
\usage{
SSGL(
  Y,
  X,
  lambda1,
  lambda0,
  groups,
  a = 1,
  b = length(unique(groups)),
  updateSigma = TRUE,
  M = 10,
  error = 0.001,
  betaStart = rep(0, dim(X)[2]),
  sigmasqStart,
  theta,
  printWarnings = TRUE,
  forceGroups = numeric()
)
}
\description{
A function to pergorm ssgl algorithm
}
