% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSGL_noSVD.R
\name{SSGL_noSVD}
\alias{SSGL_noSVD}
\title{A function to pergorm fast ssgl algorithm}
\usage{
SSGL_noSVD(
  Y,
  X,
  Xstar,
  Xtilde,
  Qmat,
  Dvec,
  lambda1,
  lambda0,
  groups,
  a = 1,
  b = length(unique(groups)),
  updateSigma = TRUE,
  M = 10,
  error = 0.001,
  betaStart = rep(0, dim(X)[2]),
  sigmasqStart,
  theta,
  printWarnings = TRUE,
  forceGroups = numeric()
)
}
\arguments{
\item{Y}{a vector of response variable}

\item{X}{a matrix of predictors}

\item{lambda0}{inital value of lambda}

\item{groups}{a vector indicting the group number}

\item{a, b}{default parameter setup}

\item{updateSigma}{T/F indicating to update sigma value}

\item{M}{default is 10}

\item{error}{numeric value for tolerance}

\item{betaStart}{vector default set to all 0}

\item{sigmasqStart}{updated value through loop}

\item{theta}{initial theta value}

\item{forceGroups}{a vector indicating index to force into same group}
}
\value{
A list of results from ssgl algorithm
}
\description{
A function to pergorm fast ssgl algorithm
}
