% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSGL.R
\name{SSGLpath}
\alias{SSGLpath}
\title{The complete path functions calls the
partial improved ssgl}
\usage{
SSGLpath(
  Y,
  X,
  lambda1,
  lambda0,
  G,
  lambda0seq = seq(lambda1, lambda0, length = 20),
  groups,
  a = 1,
  b = length(unique(groups)),
  M = 10,
  error = 0.001,
  forceGroups = c()
)
}
\description{
The complete path functions calls the
partial improved ssgl
}
