% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSGLpath_onceSVD.R
\name{SSGLpath_onceSVD}
\alias{SSGLpath_onceSVD}
\title{### C++ version with SVD adjusted
A function to pergorm fast ssgl algorithm}
\usage{
SSGLpath_onceSVD(
  Y,
  X,
  lambda1,
  lambda0,
  G,
  lambda0seq = seq(lambda1, lambda0, length = 20),
  groups,
  a = 1,
  b = length(unique(groups)),
  M = 10,
  error = 0.001,
  forceGroups = c()
)
}
\arguments{
\item{Y}{a vector of response variable}

\item{X}{a matrix of predictors}

\item{lambda1}{inital value of lambda1}

\item{lambda0}{inital value of lambda0}

\item{G}{integer value}

\item{lambda0seq}{sequence from lambda0 to lambda1}

\item{a, b}{default parameter setup}

\item{M}{default is 10}

\item{error}{numeric value for tolerance}

\item{forceGroups}{a vector indicating index to force into same group}
}
\value{
A list of results from ssgl path algorithm
}
\description{
### C++ version with SVD adjusted
A function to pergorm fast ssgl algorithm
}
